#1 - Expand crypto.c into multiple files for better tracking, but exposing functions;
